requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via
    #   -r requirements.in
    #   django
certifi==2021.10.8
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==2.0.12
    # via
    #   -r requirements.in
    #   requests
click==8.0.3
    # via
    #   -r requirements.in
    #   pip-tools
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
django==3.2.9
    # via
    #   -r requirements.in
    #   djangorestframework
django-environ==0.8.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
idna==3.3
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via -r requirements.in
itypes==1.2.0
    # via
    #   -r requirements.in
    #   coreapi
jinja2==3.0.3
    # via
    #   -r requirements.in
    #   coreschema
markupsafe==2.0.1
    # via
    #   -r requirements.in
    #   jinja2
packaging==21.3
    # via -r requirements.in
pep517==0.12.0
    # via
    #   -r requirements.in
    #   pip-tools
pip-tools==6.5.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pyparsing==3.0.7
    # via
    #   -r requirements.in
    #   packaging
pytz==2021.3
    # via
    #   -r requirements.in
    #   django
    #   djangorestframework
requests==2.27.1
    # via
    #   -r requirements.in
    #   coreapi
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via
    #   -r requirements.in
    #   ruamel-yaml
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   -r requirements.in
    #   pep517
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==1.26.8
    # via
    #   -r requirements.in
    #   requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
